requirements.txt
ops
pyOpenSSL
pyjks
netifaces
git+https://github.com/juju/charm-helpers.git#egg=charmhelpers
